Fix bug in deleteOldObservations
lat and long passed to ephem() must either be radians as a float, or degrees as a string eg 0.91 or "51.8". 
By passing degrees as a float, it is being interpreted as a value in radians. This is then normalised to the range -pi/2 to pi/2 so for example 51.8 converts to 1.5345. equal to 87.9 degrees north. 
The subsequent call to work out sunrise is then failing for half the year since the sun never rises at this latitude during the Winter. As a result, deleteOldObservations is silently failing and no space is freed.