Update GenerateMP4s.py
Patch to pass bare filename of FTPdetectinfo file to GenerateMP4s, to keep consistency with previous behaviour and avoid a crash if dir_path is relative.  If dir_path is a relative path, it gets prepended to the ftpfile_name in loadFTPdetectinfo(), and so if the latter already contains a path, then the path is prepended twice!