reinstate console output
By design this standalone module needs to send output to the console.  Unfortunately a change to the logger suppresses output unless stdout=True is added to the call to getLogger(). This PR adds that
ps I'm unsure when this bug was introduced, but my stations were working correctly on 26/07 but broken by 28/07 so i would guess the issue was introduced between those dates, possibly by merging an older PR.